perf(cli): speed up update startup and planning
