+++ src/cli/update-cli/update-command.ts
@@ -21,10 +21,10 @@ import {
 } from "../../infra/update-channels.js";
 import { fetchNpmPackageTargetStatus } from "../../infra/update-check-package-target.js";
 import {
-  checkUpdateStatus,
   compareSemverStrings,
   resolveExtendedStablePackage,
   resolveNpmChannelTag,
+  resolveUpdateInstallKind,
 } from "../../infra/update-check.js";
 import { readControlPlaneUpdateSentinelMeta } from "../../infra/update-control-plane-sentinel.js";
 import {
@@ -42,6 +42,10 @@ import {
 } from "../../infra/update-global.js";
 import { updateInstallRootsMatch } from "../../infra/update-install-root.js";
 import { cleanupStaleManagedServiceUpdateHandoffs } from "../../infra/update-managed-service-handoff-cleanup.js";
+import {
+  POST_CORE_UPDATE_CHANNEL_ENV,
+  POST_CORE_UPDATE_ENV,
+} from "../../infra/update-post-core-context.js";
 import { loadInstalledPluginIndexInstallRecords } from "../../plugins/installed-plugin-index-records.js";
 import { defaultRuntime } from "../../runtime.js";
 import type { OpenClawSchemaVersions } from "../../state/openclaw-schema-versions.js";
@@ -50,6 +54,11 @@ import { assertOpenClawStateWriteAllowedAtPath } from "../../state/openclaw-stat
 import { VERSION } from "../../version.js";
 import { resolveCliName } from "../cli-name.js";
 import { createUpdateProgress } from "./progress.js";
+import {
+  checkTargetDatabaseSchemas,
+  formatSchemaRefusalLines,
+  hasSchemaRefusal,
+} from "./schema-preflight.js";
 import {
   DEFAULT_PACKAGE_NAME,
   createGlobalCommandRunner,
@@ -61,35 +70,21 @@ import {
   resolveNodeRunner,
   resolveTargetVersion,
   resolveUpdateRoot,
+  tryResolveInvocationCwd,
   type UpdateCommandOptions,
 } from "./shared.js";
 import { suppressDeprecations } from "./suppress-deprecations.js";
 import { maybeRepairLegacyConfigForUpdateChannel } from "./update-command-config.js";
 import { printUpdateDryRun } from "./update-command-dry-run.js";
-import { executeMutableUpdate } from "./update-command-execution.js";
-import {
-  checkTargetDatabaseSchemas,
-  formatSchemaRefusalLines,
-  hasSchemaRefusal,
-} from "./update-command-git.js";
-import {
-  POST_CORE_UPDATE_CHANNEL_ENV,
-  POST_CORE_UPDATE_ENV,
-  reportPreMutationUpdateFailure,
-} from "./update-command-post-core.js";
-import { finishUpdate } from "./update-command-post-update.js";
-import { resumePostCoreUpdate } from "./update-command-resume.js";
+import { reportPreMutationUpdateFailure } from "./update-command-result.js";
 import { resolveServiceRefreshEnv } from "./update-command-service-env.js";
 import {
   gatewayServiceCommandUsesRoot,
-  resolveManagedServiceNodeRunnerOverride,
-  resolveManagedServicePackageUpdateRoot,
+  resolveManagedServicePackageUpdatePlan,
   resolvePackageRuntimePreflight,
-  tryResolveInvocationCwd,
   type ManagedServiceRootRedirect,
-  type UpdateCommandRecoveryState,
-} from "./update-command-service.js";
-export { updateFinalizeCommand } from "./update-command-finalize.js";
+} from "./update-command-service-plan.js";
+import type { UpdateCommandRecoveryState } from "./update-command-service.js";
 
 const CLI_NAME = resolveCliName();
 const DEFAULT_UPDATE_STEP_TIMEOUT_MS = 30 * 60_000;
@@ -218,6 +213,7 @@ async function updateCommandInternal(
 
   let root = discoveredRoot;
   if (postCoreUpdateResume) {
+    const { resumePostCoreUpdate } = await import("./update-execution.runtime.js");
     await resumePostCoreUpdate({
       root,
       channel: postCoreUpdateChannel,
@@ -230,17 +226,12 @@ async function updateCommandInternal(
   if (!opts.json) {
     defaultRuntime.log(theme.muted("Checking for updates..."));
   }
-  const updateStatus = await checkUpdateStatus({
-    root,
-    timeoutMs: timeoutMs ?? 3500,
-    fetchGit: false,
-    includeRegistry: false,
-  });
+  const installKind = await resolveUpdateInstallKind(root);
 
-  if (requestedChannel === "extended-stable" && updateStatus.installKind === "git") {
+  if (requestedChannel === "extended-stable" && installKind === "git") {
     await reportPreMutationUpdateFailure({
       root,
-      installKind: updateStatus.installKind,
+      installKind,
       reason: "unsupported_git_channel",
       opts,
       controlPlaneUpdateSentinelMeta,
@@ -269,7 +260,6 @@ async function updateCommandInternal(
     return;
   }
 
-  const installKind = updateStatus.installKind;
   const channel =
     requestedChannel ??
     storedChannel ??
@@ -338,17 +328,16 @@ async function updateCommandInternal(
   let packageTargetSchemaVersions: OpenClawSchemaVersions | undefined;
   let packageRuntimeTarget: { version: string; nodeEngine: string | null } | undefined;
   let managedServiceRootRedirect: ManagedServiceRootRedirect | null = null;
-  // Resolved independently of the root redirect so it covers the common case
-  // where the package root is the same but the user's PATH-resolved node
-  // differs from the node baked into the managed gateway service unit.
+  // The service's Node can differ even when its package root matches the shell.
   let managedServiceNodeRunner: string | undefined;
   let packageUpdateNodeRunner: string | undefined;
 
   if (updateInstallKind === "package") {
-    managedServiceRootRedirect = await resolveManagedServicePackageUpdateRoot({ root });
+    const servicePlan = await resolveManagedServicePackageUpdatePlan({ root });
+    managedServiceRootRedirect = servicePlan.rootRedirect;
+    managedServiceNodeRunner = servicePlan.nodeRunner;
     if (managedServiceRootRedirect) {
       root = managedServiceRootRedirect.root;
-      managedServiceNodeRunner = managedServiceRootRedirect.nodeRunner;
       if (!opts.json) {
         defaultRuntime.log(
           theme.muted(
@@ -371,22 +360,17 @@ async function updateCommandInternal(
           );
         }
       }
-    } else {
-      // Roots match but the node binary may still differ (e.g. user switched
-      // nvm/fnm/brew node after gateway install).
-      managedServiceNodeRunner = await resolveManagedServiceNodeRunnerOverride();
-      if (managedServiceNodeRunner && !opts.json) {
-        defaultRuntime.log(
-          theme.warn(
-            `Current Node (${resolveNodeRunner()}) differs from the managed gateway service Node (${managedServiceNodeRunner}).`,
-          ),
-        );
-        defaultRuntime.log(
-          theme.muted(
-            `Using the managed service Node for this update so the gateway can start after the upgrade.`,
-          ),
-        );
-      }
+    } else if (managedServiceNodeRunner && !opts.json) {
+      defaultRuntime.log(
+        theme.warn(
+          `Current Node (${resolveNodeRunner()}) differs from the managed gateway service Node (${managedServiceNodeRunner}).`,
+        ),
+      );
+      defaultRuntime.log(
+        theme.muted(
+          `Using the managed service Node for this update so the gateway can start after the upgrade.`,
+        ),
+      );
     }
     packageUpdateNodeRunner = managedServiceNodeRunner;
   }
@@ -627,6 +611,9 @@ async function updateCommandInternal(
       );
     }
   }
+
+  // Preload execution and recovery before the package swap can remove these chunks.
+  const { executeMutableUpdate, finishUpdate } = await import("./update-execution.runtime.js");
 
   // Startup migrations belong to the freshly installed Doctor. Admit shared-state
   // mutation only after every pre-install refusal has passed.
