+++ src/cli/update-cli/update-execution.runtime.ts
@@ -1,0 +1,4 @@
+// Planning stays light; load this whole surface before replacing the installed package.
+export { executeMutableUpdate } from "./update-command-execution.js";
+export { finishUpdate } from "./update-command-post-update.js";
+export { resumePostCoreUpdate } from "./update-command-resume.js";
