+++ src/cli/update-finalization-output.test-support.ts
@@ -32,12 +32,6 @@ export const readConfigFileSnapshot = async () => ({ valid: true, config, source
 export const assertConfigWriteAllowedInCurrentMode = () => {};
 `;
 const stubs = new Map<string, string>([
-  [
-    sourceUrl("./update-cli/update-command.ts"),
-    `export { updateFinalizeCommand } from ${JSON.stringify(sourceUrl("./update-cli/update-command-finalize.ts"))}; export const updateCommand = async () => { throw new Error('Unexpected core update'); };`,
-  ],
-  [sourceUrl("./update-cli/status.ts"), "export const updateStatusCommand = async () => {};"],
-  [sourceUrl("./update-cli/wizard.ts"), "export const updateWizardCommand = async () => {};"],
   [sourceUrl("../commands/doctor.ts"), doctorSource],
   [sourceUrl("../config/config.ts"), snapshotSource],
   [
@@ -73,11 +67,11 @@ const stubs = new Map<string, string>([
     `export const updatePluginsAfterCoreUpdate = async ({opts}) => { if (opts.restart !== false) throw new Error('Unexpected restart'); return {status: ${JSON.stringify(scenario === "plugin-error" ? "error" : "ok")}, changed: false}; };`,
   ],
   [
-    sourceUrl("./update-cli/update-command-post-core.ts"),
+    sourceUrl("./update-cli/update-command-result.ts"),
     "export const reportPreMutationUpdateFailure = async () => { throw new Error('Unexpected channel'); };",
   ],
   [
-    sourceUrl("./update-cli/update-command-service.ts"),
+    sourceUrl("./update-cli/update-command-service-env.ts"),
     "export const stripGatewayServiceMarkerEnv = env => ({...env}); export const disableUpdatedPackageCompileCacheEnv = env => ({...env});",
   ],
   [
