+++ src/infra/update-post-core-context.ts
@@ -3,6 +3,11 @@ import { mergeProcessEnv } from "./process-env.js";
 import type { UpdateChannel } from "./update-channels.js";
 
 export const POST_CORE_UPDATE_ENV = "OPENCLAW_UPDATE_POST_CORE";
+export const POST_CORE_UPDATE_CHANNEL_ENV = "OPENCLAW_UPDATE_POST_CORE_CHANNEL";
+export const POST_CORE_UPDATE_RESULT_PATH_ENV = "OPENCLAW_UPDATE_POST_CORE_RESULT_PATH";
+export const POST_CORE_UPDATE_INSTALL_RECORDS_PATH_ENV =
+  "OPENCLAW_UPDATE_POST_CORE_INSTALL_RECORDS_PATH";
+export const POST_CORE_UPDATE_STARTED_AT_ENV = "OPENCLAW_UPDATE_POST_CORE_STARTED_AT_MS";
 export const POST_CORE_UPDATE_REQUESTED_CHANNEL_ENV = "OPENCLAW_UPDATE_POST_CORE_REQUESTED_CHANNEL";
 export const POST_CORE_UPDATE_SOURCE_CONFIG_PATH_ENV =
   "OPENCLAW_UPDATE_POST_CORE_SOURCE_CONFIG_PATH";
